requirements.txt
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Pillow==10.0.1
SQLAlchemy==2.0.21
WTForms==3.0.1
Werkzeug==2.3.7
alembic==1.12.0
blinker==1.6.2
email-validator==2.0.0
eventlet==0.33.3
gunicorn==21.2.0
itsdangerous==2.1.2
python-dotenv==1.0.0